Forgot to commit yesterday

+++ utils/files_test.go
@@ -2,12 +2,13 @@ package utils
 
 import (
 	"os"
+	"path"
 	"testing"
 )
 
 func TestCheckAndCreateFolder(t *testing.T) {
 	var err error
-	testfolder := "testfolder"
+	testfolder := "test"
 	if err = CheckAndCreateFolder(testfolder); err != nil {
 		t.Error("Unable to create folder")
 	}
@@ -16,6 +17,9 @@ func TestCheckAndCreateFolder(t *testing.T) {
 	}
 	if err = CheckAndCreateFolder(testfolder); err != nil {
 		t.Error("Error when folder already exists")
+	}
+	if err = CheckAndCreateFolder(path.Join("/root/", testfolder)); err == nil {
+		t.Error("Should return a permission error when creating in /root/")
 	}
 	if err = os.Remove(testfolder); err != nil {
 		t.Error("Could not remove test folder")
